Change all metrics classes and folders structure
