+++ pycktool/metrics/method_metric.py
@@ -1,0 +1,13 @@
+from abc import abstractmethod
+
+from pycktool.metrics.base_metric import BaseMetric
+
+class MethodMetric(BaseMetric):
+    
+    @abstractmethod
+    def calculate(self, method_obj, context: dict = None):
+        """
+        Calculate the metric given a Method object.
+        The optional context can include additional information if needed.
+        """
+        pass
